refac: addjust the namespace of solutions
Changed as per review request
