+++ src/Exceptions/Solutions/GroupDoesNotExistSolution.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace Junges\ACL\Solutions;
+namespace Junges\ACL\Exceptions\Solutions;
 
 use Facade\IgnitionContracts\RunnableSolution;
-use Junges\ACL\Exceptions\PermissionAlreadyExistsException;
+use Junges\ACL\Exceptions\GroupAlreadyExistsException;
 
-class PermissionDoesNotExistSolution implements RunnableSolution
+class GroupDoesNotExistSolution implements RunnableSolution
 {
     /**
-     * The slug to build the permission off of.
+     * The slug to build the group off of.
      *
      * @var string
      */
@@ -21,12 +21,12 @@ class PermissionDoesNotExistSolution implements RunnableSolution
 
     public function getSolutionTitle(): string
     {
-        return 'The Permission does not exist';
+        return 'The Group '.$this->slug.' does not exist';
     }
 
     public function getSolutionDescription(): string
     {
-        return 'Did you forget to create the permission `'.$this->slug.'` with `php artisan permission:create`?';
+        return 'Did you forget to create the group `'.$this->slug.'` with `php artisan group:create`?';
     }
 
     public function getDocumentationLinks(): array
@@ -36,26 +36,26 @@ class PermissionDoesNotExistSolution implements RunnableSolution
 
     public function getSolutionActionDescription(): string
     {
-        return 'Pressing the button below will try to create the missing permission for you.';
+        return 'Pressing the button below will try to create the missing group for you.';
     }
 
     public function getRunButtonText(): string
     {
-        return 'Create Permission';
+        return 'Create Group';
     }
 
     public function run(array $parameters = [])
     {
-        $permissionModel = app(config('acl.models.permission'));
+        $groupModel = app(config('acl.models.group'));
+        $group = $groupModel->where('slug', $parameters['slug'])
+                            ->orWhere('name', $parameters['name'])
+                            ->first();
 
-        $permission = $permissionModel->where('slug', $parameters['slug'])
-                                        ->orWhere('name', $parameters['name'])
-                                        ->first();
-        if (! is_null($permission)) {
-            throw PermissionAlreadyExistsException::create();
+        if (! is_null($group)) {
+            throw GroupAlreadyExistsException::create();
         }
 
-        $permissionModel->create($parameters);
+        $groupModel->create($parameters);
     }
 
     /*
@@ -71,7 +71,7 @@ class PermissionDoesNotExistSolution implements RunnableSolution
         return [
             'name' => $name,
             'slug' => $this->slug,
-            'description' => $name.' permission',
+            'description' => $name.' group',
         ];
     }
 }
